Don't display preview link for invalid addresses
We don't want users to preview notifications we know will fail if sent to Notify. It might give the impression it will succeed when we know it won't.

So, we move the existing logic along with our new logic into a new function. This also means we can now simply determine whether to show the link based on whether it is populated, rather than via an additional flag.