Update check presenter tests for preview link
We came to the tests to amend them to ensure our logic for the preview link was covered. We also needed to drop the 'display' flag.

However, we ended up also doing some housekeeping. So, apologies that are more extensive than expected.

In the housekeeping we've moved the preview tests so that they are in the context of the `recipients:` property. We retain the different journey and notice types, though now we ensure _all_ are covered.

We've dropped the 'formatting' test. It was duplicating what had already been covered in the test above, and is not a 'property' anyway.

There also seemed to be a mix of ways to obtain the recipient we wished to use within a test. We've opted for the simplest of the methods on other. We felt it was easier for a reader to relate the recipient being asserted using the index and comparing that to the top level test that lists all the recipients.