Fix dynamic abstraction alert heading https://eaflood.atlassian.net/browse/WATER-5133 It has been spotted during QA that the dynamic heading relating to "restriction type and threshold" on the "Check the licence matches for the selected thresholds" and "Check the recipient previews" pages is not working as expected.
If the monitoring station has both "Flow" and "Level" thresholds but only the "Flow" thresholds were selected we would expect the heading to be "Flow restriction type and threshold". However in this case the heading "Flow and level restriction type and threshold" is incorrectly displayed.

This issue will be fixed in this PR by setting the heading based on the data being displayed rather than the data associated with the monitoring station.