Fix dynamic abstraction alert heading (#2205)
https://eaflood.atlassian.net/browse/WATER-5133
It has been spotted during QA that the dynamic heading relating to "restriction type and threshold" on the "Check the licence matches for the selected thresholds" and "Check the recipient previews" pages is not working as expected.

If the monitoring station has both "Flow" and "Level" thresholds but only the "Flow" thresholds were selected we would expect the heading to be "Flow restriction type and threshold". 

<img width="843" height="367" alt="Screenshot 2025-07-18 at 11 46 21" src="https://github.com/user-attachments/assets/015458e3-5750-44fc-bf09-780e851c8570" />

However in this case the heading "Flow and level restriction type and threshold" is incorrectly displayed.

<img width="994" height="301" alt="Screenshot 2025-07-18 at 11 46 38" src="https://github.com/user-attachments/assets/d3e20047-c449-498c-a502-bf764dc95398" />

<img width="994" height="219" alt="Screenshot 2025-07-18 at 11 47 07" src="https://github.com/user-attachments/assets/705ed59b-8df9-414a-aa1a-9a2b7ee2d995" />

This issue will be fixed in this PR by setting the heading based on the data being displayed rather than the data associated with the monitoring station.