Ad-hoc notice contact type view page (#2214)
https://eaflood.atlassian.net/browse/WATER-5043

When a licence version has been changed either by conditions or someone else taking it over we some times need to send out ad hoc notifications to users. This page is part of the journey to create those notification and send them.

This PR sets up the select a recipient contact type for viewing page.