Update notices setup recipients (#2222)
* Update notices setup recipients

https://eaflood.atlassian.net/browse/WATER-5041

This change adds a new 'RecipientsService' to handle the fetching and de duplication of the recipients.

Future changes will need to use this logic, so it has been deemed suitable extra this logic into a single service.

As part of this refactor we have identified the 'contacts' will be need in coming changes and so has been updated as part of this change.

* chore: pre pr checks

* chore: pre pr checks

* chore: pre pr checks