Update notices setup select recipients to fetch recipients (#2221)
* Update notices setup select recipients to fetch recipients

https://eaflood.atlassian.net/browse/WATER-5041

This change adds the recipients to the page. The recipients are transformed into to contacts and displayed to the user.

This uses the common functionality from the 'RecipientsService'

This change also ensures the validation is working.

The checkboxes are supposed to be defaulted to checked. The logic  to submit unchecked boxes will be added in another change.