Update notices setup select recipients to handle selected recipients
https://eaflood.atlassian.net/browse/WATER-5041

This change adds the selected recipients to the session and marks them as checked in the presenter.

All recipients are initially set to 'checked'. When the user has unselected any recipients they will not be 'checked' until updated by the user again.

This change does not update the common service to remove the unselected recipients from the relevant recipients. This will be done in another change.