Select how to contact recipient submit logic (#2225)
https://eaflood.atlassian.net/browse/WATER-5043

As part of the ad-hoc journey we need to be able to select which type of contact method to use to send the notice.

This PR adds the logic to submit that choice. If it is by post we save the recipient name and progress to the postcode lookup page. If they choose to send it by email we take the email address, create an md5 hash of it and save it in the additionalRecipients array in the session so it is in the same format as the recipients retrieved from the database.