Fix communications notifications (#2226)
Fix communications notifications

https://eaflood.atlassian.net/browse/WATER-5188

There is a bug introduced by the new notices work we have done. We recently migrated the legacy notifications logic into 'system'.

When the legacy notifications where sent they used a delayed sending mechanism which used a 'sendAfter' date to track when a notification was sent.

When we migrated this logic we did not use this delayed mechanism (or an overly complicated message queue system). We instead send the notification when it has been submitted and save it in the database with the created date set (legacy notification do not have a created at date).

So we have two dates used to show notification sent.

This change updates the logic to use both of these dates. We determined that a 'sendAfter' should be used in place of the 'createdAt' date when showing the date.

And as the new way of sending notifications use the 'createdAt' we sort by 'createdAt' then 'sendAfter' to show the user the latest notifications (we have handled null dates in the updated query).

Another change required was updating the communications tab to use these two dates instead of the 'event.createdAt' date. This was a carry over from the legacy code.