https://eaflood.atlassian.net/browse/WATER-5041
On the 'adhoc' journey users can select the recipients to send the notice to. When we have this 'selectedRecipients' array, we only want to show and send notice to these selected recipients.

However, we still need to render the complete recipients on the select recipients page.

This change adds the logic for `adhoc` journey to handle selected or all recipients.