Update checks when triggering a bill run (#2235)
https://eaflood.atlassian.net/browse/WATER-5166

Currently where the annual TPT bill runs (2024/2025) are in progress we are blocked from running supplementary billing in the same region.

This is because when we run the checks for both standard supplementary, and two-part tariff supplementary we are just checking if any bill runs are in progress.

In this PR we are going to make this check a little more intelligent and tweak our current checks as follows to prevent this blocking from occurring:

- When creating a standard supplementary bill run, check for any standard “in progress” bill runs. Ignore any “in progress” 2PT.
- When creating a 2PT supplementary bill run, check for any 2PT “in progress” bill runs. Ignore any “in progress” standard.

The billing engines ignore each other's transactions when generating bills, so we believe there isn't any risk associated with this change.