Don't display £0.00 value for SROC 2PT sup bills in review (#2240)
https://eaflood.atlassian.net/browse/WATER-5161

SROC two-part tariff supplementary bill runs are showing a £0.00 value in the bill runs screen when they are in review. This is because the value of two-part tariff bill runs is not calculated until after they have been reviewed.

To mitigate this issue with non-supplementary SROC two-part tariff bill runs. We do not display a value at all while the bill run is in review. However this rule is not currently applied to SROC two-part tariff supplementary bill runs.

In this PR the same logic that is used for the bill run value of non-supplementary SROC two-part tariff bill runs, is going to be applied to SROC two-part tariff supplementary bill runs.