Show all purposes in return log pages (#2215)
https://eaflood.atlassian.net/browse/WATER-5150

A user pointed out that we're not displaying the full list of purposes in two places but did point out where it was displayed correctly. Having dug into the code and looking at where else we show the purposes I noticed we're doing it in two places already so I've pulled the code for it out of both of those places, combined it so it includes the break clause for being handed an empty array and then started calling it from the various places it needs to be seen so we are consistent across the site.