Update notices setup recipients to use additional recipients
https://eaflood.atlassian.net/browse/WATER-5041

When the user has created an additional recipient, we need to include these additional recipients (could be many) in the shared/ common recipients' logic.

This change adds the 'additionalRecipients' array to the RecipientsService. Regardless of other consideration the 'additionalRecipients' are always added to the recipients if present.

When adding additional recipients, we need to handle make sure the selected recipients are updated. We do this in the contact types submit handler to simplify the logic. When a new recipient is added, we add the contact hash id to the selected recipients array.

To make this work, we need to create the selected recipients on the check page. We do this incandescently to keep the code flexible (if we allow the 'manage recipients' flow on all journeys).

we have also changed the logic when a new recipient is added, we redirect them to the check page and show a notification, telling the user a new recipient has been added.

This change ensures our standard recipients array (de duplicated / selected) can be used / reliable in all areas of the code.