Add user_data column to users view (#2234)
https://eaflood.atlassian.net/browse/WATER-5171

The contact information a.k.a. profile details are stored in the JSON in the `user_data` column so this needs adding to the `users` view to be visible to the new system.

This will then allow the above ticket to be progressed, which involves moving the profile details into the new system.

Some tests also needed to be updated to ignore the new field in the model.