Upgrade `bcryptjs` Whilst the package now exports an ECMAScript module by default, it still works fine with our service without having to make any changes due to the UMD fallback.
**Breaking changes**
- Modernize project structure
The project now exports an ECMAScript module by default, albeit with an UMD fallback, ships with types, the dist/ directory no longer exists in version control, and Closure Compiler externs have been removed.

- Generate 2b hashes by default
This library was not affected by the bug that led to incrementing the bcrypt version from 2a to 2b, but nowadays most implementations use 2b, including the native bcrypt binding, so this change aligns with them. Existing hashes will continue to work, but test logic that generates hashes and compares them literally might need to be updated to account for the new default.