Bump package major versions (#2244)
https://eaflood.atlassian.net/browse/WATER-5191

Before adding **Snyk** to the CI workflow we are going to ensure that the packages are as up to date as they can be.

When doing major version upgrades we check the release notes for each package to see what breaking changes there are. If the breaking changes don't require any updates to the code (i.e. changes to the minimum Node version) these will be upgraded. If code changes are required these will be upgraded individually in separate PRs.