Reinstall `notifications-node-client` (#2248)
The package `notifications-node-client` has a nested dependency `axios`. The version of `axios` that is currently in our `package-lock.json` is v1.8.2. This version of `axios` has the issue `Server-side Request Forgery (SSRF)`.

As the package `notifications-node-client` has `axios` set to update minor versions in it's `package.json` file. Reinstalling the `notifications-node-client` package will bump it's nested dependency `axios` to the latest version which will resolve the SSRF issue.