Add basic notice setup return form PDF (#2192)
* Add basic notice setup return form PDF

https://eaflood.atlassian.net/browse/WATER-5162

This changes uses the previously implemented 'GenerateReturnFormRequest' to convert HTML into a PDF.

This change adds a new route to allow users to view the generated PDF, this is not only useful for testing, but we use this on the notice setup check preview link to allow users to preview the letter before sending. (This uses the browser's built-in PDF viewer.)

We have created the PrepareReturnFormsService to handle orchestrating fetching the data and returning the file (ArrayBuffer) to a calling service. We have implemented the preview page as part of this change.

We will need to add another service to call the PrepareReturnFormsService when se send the PDF to notify.