Switch to using Snyk CLI action in CI (#2250)
https://eaflood.atlassian.net/browse/WATER-5191

With the adoption of SSO in the Defra GitHub org, our current [Snyk](https://snyk.io) status checks have stopped working.

The [team issue](https://github.com/DEFRA/water-abstraction-team/issues/144) goes into more detail, but in essence, whilst the GitHub account that **Snyk** uses to connect has SSO enabled, it won't connect to GitHub to update the PR with the test result.

This change performs the same check without **Snyk** having to authenticate to GitHub. Instead, GitHub will authenticate with **Snyk** by using its CLI via a GitHub action.

It does mean **Snyk** will no longer be listed as a separate check; instead, it will just be a new step in our CI. So, the outcome (adding a package with a vulnerability will cause the PR to fail checks) will be the same.

Some additional info on the `form-data` package that cannot be updated:
```
  ✗ Predictable Value Range from Previous Values [Critical Severity][https://security.snyk.io/vuln/SNYK-JS-FORMDATA-10841150] in form-data@2.3.3
    introduced by @airbrake/node@2.1.9 > @airbrake/browser@2.1.9 > @types/request@2.48.8 > form-data@2.5.3 and 1 other path(s)
  This issue was fixed in versions: 2.5.4, 3.0.4, 4.0.4
```