Update notice setup return form PDF (#2246)
https://eaflood.atlassian.net/browse/WATER-5162

This change adds styling and structure to the upcoming return forms PDF work, with some mock data to render in the view. This makes it easier to swap out the legacy codes structure. A change will be made in the future to fetch this data.

This change has copied the CSS from the legacy service. This was proven to work with Gotenberg in a previous spike.

To ensure everything looks and works across multiple pages, the guidance page has been added to this change. This is just HTML with no additional data needed for the page.

The header and footer have been added to test the layout works as expected, correct pagination has not been implemented and is mocked/ hard coded for now.