Do not run snyk check in CI for dependabot
https://eaflood.atlassian.net/browse/WATER-5191

We recently updated how we use [Snyk](https://snyk.io/) to prevent packages with vulnerabilities from being added to the project. Previously, we had **Snyk** as a required check, connecting to the repo when it saw changes made to the repo.

For reasons outlined in [this team issue](https://github.com/DEFRA/water-abstraction-team/issues/144), we needed to change it to instead [run the check from GitHub as part of our CI workflow](https://github.com/DEFRA/water-abstraction-system/pull/2250).

This method relies on access to a secret token stored in the repo settings. Unfortunately, as an external party Dependabot is not permitted access to these secret tokens. This is the same issue we spotted with SonarQube and were never able to resolve.

So, this change applies the same clause we use for **SonarQube** to the **Snyk** step in our CI.