Fix error highlighting in `period-used.njk`
The `period-used` page is on our radar for fixing because it uses the `+ errorClass` method of adding the error highlighting class which we are moving away from. On investigation we realise that the highlighting isn't working anyway -- the individual day/month/year boxes aren't highlighted if there's a validation error. Moving to our preferred `(' govuk-input--error' if error)` method fixes this.