Fix bill run check page back link (#2258)
https://eaflood.atlassian.net/browse/WATER-5193

When on the check page for a two-part tariff bill run that is for financial year 2025. The back button takes you back to the "Select the season" page instead of the "Select the financial year" page.

This will be fixed in this PR by amending the logic that determines the backlink.