Add all pages for the return forms PDF
https://eaflood.atlassian.net/browse/WATER-5162

This change adds the remaining pages required for the return forms PDF.

The legacy code was using macros, and partials these have been removed in place of inline HTML. The HTML semantics were incorrect and have been updated / fixed as part of this change.

The PDF still looks as expected.