Update recipients download to use Notify addr pres (#2263)
https://eaflood.atlassian.net/browse/WATER-5147

We have a known issue with the addresses we are generating and sending to [GOV.UK Notify](https://www.notifications.service.gov.uk/).

If a contact has too many address fields populated, it results in us generating an address with more than seven lines, which Notify rejects.

We also want to verify that international addresses are handled correctly and explore various scenarios based on the address data we're finding.

We [added a Notify address presenter](https://github.com/DEFRA/water-abstraction-system/pull/2194) then [linked View and Preview recipients pages](https://github.com/DEFRA/water-abstraction-system/pull/2200) to it.

Still to update is the 'Download recipients' feature of the check recipients page. It is still using the old logic to generate address information, which means it is no longer consistent with what we are sending to Notify or showing in the UI.

This change updates it to also use our new Notify Address presenter.