Link Abs Alert preview to NotifyAddressPresenter (#2264)
https://eaflood.atlassian.net/browse/WATER-5146
https://eaflood.atlassian.net/browse/WATER-5147

In [Add notify address presenter for all scenarios](https://github.com/DEFRA/water-abstraction-system/pull/2194), we added a new presenter to ensure we do a better job of generating addresses for letters sent via Notify.

Previously, we would generate addresses that were too long (more than 7 lines) or were not properly formatted for international addresses (country last _after_ postcode if present).

Now that it is available, we need to utilise it. In [Link View and Preview to NotifyAddressPresenter](https://github.com/DEFRA/water-abstraction-system/pull/2200), we ensured the return invitation and reminder notice setup journeys used it when viewing the recipients or previewing a notification. We also added a warning to the 'check recipients' page if an invalid address has been found.

The 'check recipients' page and the warning will apply to abstraction alerts as well. This change updates the abstraction alert preview presenter to use the new `NotifyAddressPresenter`.