Ad-hoc notice - select address (#2245)
https://eaflood.atlassian.net/browse/WATER-5045

As part of sending out notifications to users we need to be able to select an address. The postcode was provided on the previous page and this page will have a list of addresses for the user to choose from.

When chosen the information will then be converted into a recipient matching the existing structure retrieved from the database. It is then added to an array to be displayed to the user on the check page. It then returns the user to that check page.