Update bill run send views to 'template' standard (#2260)
https://eaflood.atlassian.net/browse/WATER-5155

We use the changes made in https://github.com/DEFRA/water-abstraction-system/pull/2186 as a model for updating the "Cancel a bill run" journey.

Going by this PR, we're looking at implementing the following principles:

### Errors

Where we see things like this:

```njk
{% if error.text.startResult %}
    {% set startResultErrorClass = " govuk-input--error" %}
    {% set startResultErrorMessage = { text: error.text.startResult } %}
{% else %}
    {% set startResultErrorClass = "" %}
    {% set startResultErrorMessage = null %}
{% endif %}

{# ... #}

{{ govukDateInput({
    id: "abstraction-period-start",
    namePrefix: "abstraction-period-start",
    errorMessage: { text: error.text.startResult } if error.text.startResult,
    items: [
        {
            classes: "govuk-input--width-2" + startResultErrorClass,
            {# ... #}
        },
        {# ... #}
    ]
}) }}
```

We can simply do this:

```njk
{{ govukDateInput({
    id: "abstraction-period-start",
    namePrefix: "abstraction-period-start",
    errorMessage: startResultErrorMessage,
    items: [
        {
            classes: "govuk-input--width-2 {{ 'govuk-input--error' if error.text.startResult }}",
            {# ... #}
        },
        {# ... #}
    ]
}) }}
```

In other words, we ditch the first block entirely that sets the error class and message, and instead set them conditionally within whatever uses them.

### Section comments

We seem to be ditching entirely section comments, eg:

```njk
{# Main heading #}

{# Back link #}
```

etc.

### `div`s

As a general point, we usually don't need to wrap things in `<div>` tags. Wrap text in `<p>` tags; individual components don't need anything wrapping them.

### `govuk-body` class

We remove the `govuk-body` class wherever we see it. So for example. if we have `<p class="govuk-body">...</p>` anywhere then we retain the `<p>` tags but remove the class.

If we have `<div class="govuk-body">` then we would remove the `<div>` entirely in line with the previous principle.

### Page headings

If every page shares a common heading, we can extract it to a suitable file in `views/macros`. As we update more and more pages, we will likely accumulate common page headings which we can re-use across journeys (for example, there is already `views/macros/licence-reference-page-heading.njk` which shows the page title with the licence ref smaller above it; this is common in our service). When we do import a page heading, ensure we have a separating line between it and the govuk component imports, ie:

```njk
{% extends 'layout.njk' %}
{% from "govuk/components/back-link/macro.njk" import govukBackLink %}
{% from "govuk/components/button/macro.njk" import govukButton %}

{% from "macros/licence-reference-page-heading.njk" import pageHeading %}
```

We insert the page heading like so:

```njk
{{ pageHeading(licenceRef, pageTitle) }}

{# or if we want to set specific styling to match existing design: #}
{{ pageHeading(licenceRef, pageTitle, 'govuk-heading-xl govuk-!-margin-bottom-3') }}

```

If the page is largely comprised of a single component (eg. a page displaying radio buttons) we set the page heading:

```njk
{% set pageHeading %}
    {{ pageHeading(licenceRef, pageTitle) }}
{% endset %}
```

We can then use it in a `fieldset`:

#### `fieldset`

When the page is largely comprised of a single component, ensure we have a `fieldset` in it, eg:

```njk
{{ govukCheckboxes({
    name: "agreementsExceptions",
    fieldset: {
        legend: {
            html: pageHeading
        }
    },
    {# ... #}
}) }}
```

We also clean up any existing `fieldset.legend`.

### `govukButton`

We consistently write it on one line, eg:

```njk
{# Incorrect: #}
{{ govukButton({
    text: "Remove",
    preventDoubleClick: true
}) }}

{# Correct: #}
{{ govukButton({ text: "Continue", preventDoubleClick: true }) }}
```

### `govuk-summary-list`

If we have a summary list, we don't need the `govuk-summary-list` class, eg:

```njk
{# Incorrect: #}
{{ govukSummaryList({
    classes: 'govuk-!-margin-bottom-2',
    rows: [
        {
            classes: 'govuk-summary-list govuk-summary-list__row--no-border',
            key: {
                text: "Reason"
            },
            value: {
                text: reason
            }
        },
        {# ... #}
    ]
}) }}

{# Correct: #}
{{ govukSummaryList({
    classes: 'govuk-!-margin-bottom-2',
    rows: [
        {
            classes: 'govuk-summary-list__row--no-border', {# Removed class #}
            key: {
                text: "Reason"
            },
            value: {
                text: reason
            }
        },
        {# ... #}
    ]
}) }}
```

### `text` with `attributes` vs `html`

If we're using `html` somewhere but only to apply an attribute, use `text` instead. For example:

```njk
{# Incorrect: #}
{{ govukSummaryList({
    {# ... #}
    items: [
        {
            html: '<span data-test="change-reason">' + 'Change' + '</span>',
            href: "/system/return-versions/setup/" + sessionId + "/start-date",
            visuallyHiddenText: "the reason for the return requirement"
        }
    ]
}) }}

{# Correct: #}
{{ govukSummaryList({
    {# ... #}
    items: [
        {
              attributes: { 'data-test': 'change-reason' },
              text: 'Change',
              href: reasonLink, {# See below for why this was also changed #}
              visuallyHiddenText: "the reason for the return requirement"
        }
    ]
}) }}
```

### Pass links to template from presenter instead of hardcoding

In the example above, we see that `href` goes from a hardcoded url `"/system/return-versions/setup/" + sessionId + "/start-date"` to `reasonLink`, which we define in the presenter and pass to the template. This also applies to:

### Backlinks

We define `backlink` in the presenter and pass to the template:

```njk
{# Incorrect: #}
{{
    govukBackLink({
        text: 'Back',
        href: '/system/bill-runs/setup/' + sessionId + '/region'
    })
}}

{# Correct: #}
{{
    govukBackLink({
        text: 'Back',
        href: backlink
    })
}}
```

### Hint text

If we have a standalone hint on a page with a single main component, see if we can add it to the component. For example:

```njk
{# Incorrect: #}
<div class="govuk-hint"> Select all that apply </div>

{{ govukCheckboxes({
    name: "purposes",
    items: checkBoxItems,
    {# ... #}
}) }}

{# Correct: #}
{{ govukCheckboxes({
    name: "purposes",
    hint: {
        text: "Select all that apply"
    },
    items: checkBoxItems,
    {# ... #}
}) }}
```

### `govuk-label` classes

If we have some `label` text then we don't also need to add a `govuk-label` class. For example:

```njk
{# Incorrect: #}

{{ govukInput({
    label: {
        text: "Enter the volume",
        classes: "govuk-label"
    },
    {# ... #}
}) }}

{# Correct: #}
{{ govukInput({
    label: {
        text: "Enter the volume"
    },
    {# ... #}
}) }}
```

(Note some other changes there but the key thing is the removal of `classes: "govuk-label--l"`)