Fix selected recipients in check recipients page (#2266)
https://eaflood.atlassian.net/browse/WATER-5041

We've recently incorporated a 'Select recipients' page as part of the notice setup journey for ad-hoc notifications. To do this, we added a new property to the setup session: `selectedRecipients`.

When you first view the 'Check recipients' page, if the property is not present, it is initialised with all current recipients found for the notice being created. In the ad-hoc journey, you can then go to the 'Select recipients' page and amend which recipients are selected. When you return to the 'Check recipients' page, seeing that `selectedRecipients` exists, it uses it to determine which recipients to display.

The issue found is when a user goes 'back' in the setup journey and makes changes that alter which recipients are applicable, for example, changing the threshold for an abstraction alert.

Coming back to the 'Check recipients' page, it is seeing `selectedRecipients` exists, so it is still only using it to determine which recipients to display. This is even though new recipients have been found.

It can also lead to nothing being displayed. If the change to the notice means the previously 'selected' recipient is no longer applicable, the ID in `selectedRecipients` matches to nothing, resulting in an empty page.

The fix is quite simple. Should a user go back in the journey, we have to reset the session, i.e. delete the `selectedRecipients` property. Then, when they come back to the 'Check recipients' page, it will re-initialise with all applicable recipients 'selected'.

> This is a short-term fix. The final notice setup solution will incorporate a 'Check details' page for all journeys. And, like our other journeys, going 'back' is not possible from this page. Instead, changes are managed through change links.