Update return forms PDF to use real return data (#2265)
* Update return forms PDF to use real return data

https://eaflood.atlassian.net/browse/WATER-5162

This change updates the return forms PDF to use data from the selected return.

This change adds the return id to the path / controller. This allows the preview to be specific to the return selected on the selected returns page. When we implement the sending of the notice, we can pass in the return id to generate the PDF per return id (and recipient if required).

We needed to update the exciting fetch logic to return the additional data and check the return passed in has been selected from the journey.

Some general renaming has been done to update a previous assumption, The main one being the removal of the 'formatId' this is guaranteed to be the same as the 'returnReference' and has been removed.

The 'regionAndArea' will be updated in another change. This has some additional complexity to big for this change.