Update postcode page after recent changes (#2268)
https://eaflood.atlassian.net/browse/WATER-5045

As part of #2245 we introduced a change from name to contactName when saving the name of recipient in an ad-hoc journey. This then broke some links on the postcode page.

This PR fixes those links and moves the international one into the presenter so that it can be included in unit tests and we will be pick up on similar issues happening again.