Fix not using company in view billing account
https://eaflood.atlassian.net/browse/WATER-5198

It has been spotted that we are displaying the wrong company name for some billing accounts. It was spotted when testing WATER-5174 (updating changed billing accounts with their customer file reference).

QA were using the change address journey to force a change, and happened to opt to change who the bills should go to. They then proceeded to create a new billing contact.

Against the record, this new company's ID is recorded in the `public.billing_addresses` `company_id` field. But when we fetch the billing account details, we are only getting the linked contact record, not the company. This results in us always displaying the primary billing account company name when the user has opted to send the bills to a different company (and its name should be displayed).

This change fixes the view to use the billing accounts 'billing address company name' if present.