Update return forms PDF to use region and nald area code (#2270)
* Update return forms PDF to use region and nald area code for the "regionAndArea"

https://eaflood.atlassian.net/browse/WATER-5162

This change updates the 'regionAndArea' page data, for the return forms PD, to use real data.

We need the region from our 'regions' table in the db, we use the 'nald_region_id' in the metadata to join the region table to the return logs and return the region display name.

When there is an 'areaCode' in the metadata we display that in addition to the region. This is logic implied by the legacy code that the 'nald.areaCode' may not exist.