Add the 'FetchLetterRecipientsService'
https://eaflood.atlassian.net/browse/WATER-5162

We have been gradually adding logic to allow the user to send return forms (paper forms).

The recipient/s of this letter will come from our existing fetch recipients logic. But we only need the recipient where it has an address (as opposed to an email).

This change copies the existing fetch recipient logic and extracts only the letter portion of the query into a new service.