Add the 'FetchLetterRecipientsService' (#2279)
* Add the 'FetchLetterRecipientsService'

https://eaflood.atlassian.net/browse/WATER-5162

When generating a return forms PDF we need to set the address in the page data to be rendered in the form. This we can get from our existing recipients' logic.

This change adds the 'FetchLetterRecipientsService' into the recipient service. This will now return only letters when the notice type is set to 'returnForms'.

We need to pass the recipient into the prepare return forms service to add the address onto the letters cover page. We use our existing address mapper to turn the recipient into an address.

We have updated the preview link to use the contact hash and return id. We need to implement an intermediary page to allow the user to select which return id they want to 'preview'. For now, we have added a 'placeHolder' into the preview link; to ensure it functionally works, we just pick the first return id from the previously selected returns.

* chore: pre pr checks