Fix permissions for setup abstraction alert notice
https://eaflood.atlassian.net/browse/WATER-5209

After deploying the [latest release](https://github.com/DEFRA/water-abstraction-system/releases/tag/v0.40.0) of the project, our smoke testing highlighted that users with the permission set 'Environment Officer' were unable to create abstraction alerts using our new journey.

When we looked, we saw that we had copy-pasted the same scope for the routes as used on the existing setup returns reminders and invitations journeys.

This is wrong. To create a water abstraction alert you must have the `hof_notifications` role. When creating a returns invitation, reminder or return-forms (ad-hoc) the user must have the `returns` role.

This change fixes the scope for the routes to allow Environment Officers to create abstraction alerts.