Update continue btn to confirm in remove tag page (#2300)
https://eaflood.atlassian.net/browse/WATER-5003

When we migrated the 'Remove licence tag' (related to monitoring stations and abstraction alerts), we built a 'check' page to allow the user to check they are removing the correct tag.

As this is the final action in the 'journey', and we are asking them to _confirm_ the action, the button text should have been `Confirm`. However, it was `Continue`, which infers the journey is not yet complete.

This small change fixes it.