Submit international address for ad-hoc notice (#2273)
https://eaflood.atlassian.net/browse/WATER-5047

As part of sending out notifications to users we need to be able to submit an international address. They can get to this page form the postcode page by clicking on a link.

When the address has been submitted it is converted into a recipient object and added to the additionalRecipient array as well as the hash added to the selectedRecipient array.

This PR refactors the way the address is handled before being passed back to the check recipients page.